Rename python package name
Having a dash (`-`) in the Python package name is not compliant to the Python module naming convention. Instead, we should use the underscore(`_`).
