+++ provider/cmd/pulumi-resource-onepassword/schema.json
@@ -43,7 +43,7 @@
             "disableUnionOutputTypes": true
         },
         "python": {
-            "packageName": "pulumi-onepassword",
+            "packageName": "pulumi_onepassword",
             "requires": {
                 "pulumi": "\u003e=3.0.0,\u003c4.0.0"
             },
