+++ sdk/python/pulumi_onepassword/config/__init__.pyi
