+++ sdk/python/pulumi_onepassword/item.py
@@ -486,7 +486,7 @@ class Item(pulumi.CustomResource):
 
         ```python
         import pulumi
-        import pulumi-onepassword as onepassword
+        import pulumi_onepassword as onepassword
 
         demo_password = onepassword.Item("demoPassword",
             vault=var["demo_vault"],
@@ -557,7 +557,7 @@ class Item(pulumi.CustomResource):
 
         ```python
         import pulumi
-        import pulumi-onepassword as onepassword
+        import pulumi_onepassword as onepassword
 
         demo_password = onepassword.Item("demoPassword",
             vault=var["demo_vault"],
