+++ sdk/python/pulumi_onepassword/pulumi-plugin.json
